+++ hutool-core/src/main/java/cn/hutool/core/bean/BeanUtil.java
@@ -789,6 +789,12 @@ public class BeanUtil {
 		if (collection.isEmpty()) {
 			return new ArrayList<>(0);
 		}
+
+		// issue#3091
+		if(ClassUtil.isBasicType(targetType) || String.class == targetType){
+			return Convert.toList(targetType, collection);
+		}
+
 		return collection.stream().map((source) -> {
 			final T target = ReflectUtil.newInstanceIfPossible(targetType);
 			copyProperties(source, target, copyOptions);