+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # 🚀Changelog
 
 -------------------------------------------------------------------------------------------------------------
-# 5.8.19.M1 (2023-05-20)
+# 5.8.19.M1 (2023-05-23)
 
 ### 🐣新特性
 * 【db    】      优化HttpRequest.toString()内容打印（issue#3072@Github）
@@ -14,6 +14,7 @@
 * 【core  】      SyncFinisher增加setExceptionHandler方法（issue#I716SX@Gitee）
 * 【core  】      FileTypeUtil.getType增加文件判断（pr#3112@Github）
 * 【core  】      增加CsvWriteConfig.setEndingLineBreak配置项（issue#I75K5G@Gitee）
+* 【core  】      增加Tailer追踪文件时文件被删除的处理情况（pr#3115@Github）
 
 ### 🐞Bug修复
 * 【core  】      修复URLUtil.decode无法解码UTF-16问题（issue#3063@Github）